+++ src/legacyEvents/ContractCall.ts
@@ -1,5 +1,5 @@
-import { LegacyTransaction } from '../legacy-transaction/LegacyTransaction'
-import { LegacyTransactionType } from '../legacy-transaction/LegacyTransactionType'
+import { LegacyTransaction } from '../legacyTransaction/LegacyTransaction'
+import { LegacyTransactionType } from '../legacyTransaction/LegacyTransactionType'
 
 export class ContractCall extends LegacyTransactionType {
   matches(transaction: LegacyTransaction): boolean {
