+++ src/legacyEvents/TokenReceived.ts
@@ -1,27 +1,29 @@
 import { LegacyEventBuilder } from '../helpers/LegacyEventBuilder'
 import { LegacyEventTypes } from '../resolvers'
-import { LegacyTransaction } from '../legacy-transaction/LegacyTransaction'
-import { LegacyTransactionType } from '../legacy-transaction/LegacyTransactionType'
+import { LegacyTransaction } from '../legacyTransaction/LegacyTransaction'
+import { LegacyTransactionType } from '../legacyTransaction/LegacyTransactionType'
 
-export class Faucet extends LegacyTransactionType {
+export class TokenReceived extends LegacyTransactionType {
   matches(transaction: LegacyTransaction): boolean {
     return (
       transaction.transfers.length === 1 &&
-      transaction.transfers.containsFaucetTransfer()
+      transaction.transfers.containsTransferTo(this.context.userAddress)
     )
   }
 
   getEvent(transaction: LegacyTransaction) {
-    const transfer = transaction.transfers.getFaucetTransfer()
+    const transfer = transaction.transfers.getTransferTo(
+      this.context.userAddress,
+    )
 
     if (!transfer) {
-      throw new Error('Transfer from faucet not found.')
+      throw new Error('Transfer to the user not found.')
     }
 
     return LegacyEventBuilder.transferEvent(
       transaction,
       transfer,
-      LegacyEventTypes.FAUCET,
+      LegacyEventTypes.RECEIVED,
       transfer.fromAddressHash,
       transfer.fromAccountHash,
     )
